requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
python-dotenv==0.20.0
requests==2.27.1
rocket-python==1.3.0
rocketchat==1.0.0
rocketchat-API==1.24.0
urllib3==1.26.9
Werkzeug==2.1.1
WTForms==3.0.1
